简化加载sft权重

+++ rlhf_ilql/train.py
@@ -105,7 +105,11 @@ if __name__ == '__main__':
     pl_model = MyILQLTransformer(config=config,model_args=model_args,training_args=training_args,lora_args=lora_args,ilql_args=ilql_args,
                                  load_in_8bit=load_in_8bit,device_map={"": trainer.fabric.local_rank} if trainer.world_size > 1 else "auto")
 
-    pl_model.bfloat16()
+    # pl_model.bfloat16()
+    pl_model.float()
+
+    # 如果自定义训练了sft_weight , 可以再次加载sft_weight
+    # pl_model.load_sft_weight('sft_weight.bin')
 
     ckpt_path = './best_ckpt/best.pt'
     if not data_args.convert_onnx:
