+++ rlhf_ppo/train.py
@@ -152,9 +152,11 @@ if __name__ == '__main__':
     pl_model = MyPPOTransformer(config=config,model_args=model_args,training_args=training_args,lora_args=lora_args,ppo_args=ppo_args,
                                 load_in_8bit=load_in_8bit,device_map={"": trainer.fabric.local_rank} if trainer.world_size > 1 else "auto")
 
-    # 如果使用  Trainer.precision = '16-mixed',
-    #pl_model.float()
-    pl_model.bfloat16()
+    # pl_model.bfloat16()
+    pl_model.float()
+
+    # 如果自定义训练了sft_weight , 可以再次加载sft_weight
+    # pl_model.load_sft_weight('sft_weight.bin')
 
     # pl_ref_model = load_ref_model('../reward/best_ckpt')
     pl_ref_model = copy.deepcopy(pl_model)
